refactor: :truck: change file structure & move some files
