+++ __tests__/user.test.js
@@ -1,4 +1,4 @@
-const app = require("../app")
+const app = require("../src/app")
 const request = require("supertest");
 require("dotenv").config();
